+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 - Include alias and materialized fields in queryset results
 - Check for database existence, to allow delayed creation
 - Added `Database.does_table_exist` method
+- Support for `IS NULL` and `IS NOT NULL` in querysets (kalombos)
 
 v1.0.1
 ------