+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Change Log
 ==========
 
-Unreleased
+v1.0.2
 ----------
 - Include alias and materialized fields in queryset results
 - Check for database existence, to allow delayed creation